requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Templates
jinja2>=3.1.0

# Data validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0

# For local Fidelity scraping (not needed on Railway)
# playwright>=1.40.0  # Uncomment for local development
